backend/production-requirements.txt
Django==4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
pillow==11.0.0
